// src/global/enseignant.d.ts
export interface Enseingnant{
    adresse : string;
    codePostal: string;
    emailPerso : string;
    emailUbo :string;
    mobile : string;
    noEnseignant : number;
    nom : string;
    pays : string;
    prenom : string;
    sexe : string;
    telephone : string;
    type : string;
    ville : string;
}